# packages/shared/generated/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "IntakeAnswerRequest": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "base_version": {
            "minimum": 0.0,
            "title": "Base Version",
            "type": "integer"
          },
          "question_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Question Id",
            "type": "string"
          },
          "skipped": {
            "default": false,
            "title": "Skipped",
            "type": "boolean"
          }
        },
        "required": [
          "question_id",
          "base_version"
        ],
        "title": "IntakeAnswerRequest",
        "type": "object"
      },
      "IntakeDraftRequest": {
        "additionalProperties": false,
        "properties": {
          "base_version": {
            "minimum": 0.0,
            "title": "Base Version",
            "type": "integer"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "base_version",
          "title"
        ],
        "title": "IntakeDraftRequest",
        "type": "object"
      },
      "IntakeDraftResponse": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "task_id",
          "status",
          "version"
        ],
        "title": "IntakeDraftResponse",
        "type": "object"
      },
      "IntakeFactSummary": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "enum": [
              "unconfirmed",
              "confirmed",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "value",
          "status"
        ],
        "title": "IntakeFactSummary",
        "type": "object"
      },
      "IntakeQuestionResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "conflict",
                  "missing_unit",
                  "ambiguous_role"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "enum": [
              "short_answer",
              "deep_answer"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "prompt"
        ],
        "title": "IntakeQuestionResponse",
        "type": "object"
      },
      "IntakeSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "answered_question_ids": {
            "items": {
              "type": "string"
            },
            "title": "Answered Question Ids",
            "type": "array"
          },
          "completed_count": {
            "title": "Completed Count",
            "type": "integer"
          },
          "current_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IntakeQuestionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "fact_summaries": {
            "items": {
              "$ref": "#/components/schemas/IntakeFactSummary"
            },
            "title": "Fact Summaries",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "remaining_estimate": {
            "title": "Remaining Estimate",
            "type": "integer"
          },
          "resume_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Id"
          },
          "skipped_question_ids": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Question Ids",
            "type": "array"
          },
          "status": {
            "enum": [
              "active",
              "drafting",
              "completed",
              "abandoned"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "version",
          "current_question",
          "completed_count",
          "remaining_estimate",
          "answered_question_ids",
          "skipped_question_ids",
          "fact_summaries",
          "task_id",
          "resume_id"
        ],
        "title": "IntakeSessionResponse",
        "type": "object"
      },
      "IntakeStartRequest": {
        "additionalProperties": false,
        "properties": {
          "restart": {
            "default": false,
            "title": "Restart",
            "type": "boolean"
          }
        },
        "title": "IntakeStartRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "AI Resume API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/intake-sessions": {
      "post": {
        "operationId": "start_intake_v1_intake_sessions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Start Intake",
        "tags": [
          "intake"
        ]
      }
    },
    "/v1/intake-sessions/{session_id}": {
      "get": {
        "operationId": "get_intake_v1_intake_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Intake",
        "tags": [
          "intake"
        ]
      }
    },
    "/v1/intake-sessions/{session_id}/answers": {
      "post": {
        "operationId": "answer_intake_v1_intake_sessions__session_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Answer Intake",
        "tags": [
          "intake"
        ]
      }
    },
    "/v1/intake-sessions/{session_id}/drafts": {
      "post": {
        "operationId": "create_draft_v1_intake_sessions__session_id__drafts_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Draft",
        "tags": [
          "intake"
        ]
      }
    },
...
  }
}